5_8_1 5_8_2 - маленький рефакторинг + галерея выбора фоток
