+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:places/mocks.dart';
 import 'package:places/shared/places_texts.dart';
 import 'package:places/shared/theme_notifier.dart';
-import 'package:places/ui/screen/add_sight_screen.dart';
+import 'package:places/ui/screen/add_sight_screen/add_sight_screen.dart';
 import 'package:places/ui/screen/filters_screen.dart';
 import 'package:places/ui/screen/res/themes.dart';
 import 'package:places/ui/screen/settings_screen.dart';
@@ -38,8 +38,8 @@ class _AppState extends State<App> {
       theme: _isDarkTheme ? darkTheme : lightTheme,
       // theme: darkTheme,
       title: PlacesTexts.appTitle,
-      // home: SightAddScreen(),
-      home: VisitingScreen(),
+      home: AddSightScreen(),
+      // home: VisitingScreen(),
       // home: SightListScreen(),
       // home: SightDetails(mocks[2]),
       // home: FiltersScreen(),
