+++ lib/ui/screen/add_sight_screen/types.dart
@@ -1,0 +1,22 @@
+class AddFormData {
+  final String name;
+  final double lat;
+  final double lon;
+  final String details;
+
+  AddFormData({
+    this.name,
+    this.lat,
+    this.lon,
+    this.details,
+  });
+
+  AddFormData update({name, lat, lon, details}) {
+    return AddFormData(
+      name: name ?? this.name,
+      lat: lat ?? this.lat,
+      lon: lon ?? this.lon,
+      details: details ?? this.details,
+    );
+  }
+}
