+++ lib/ui/screen/sight_list_screen.dart
@@ -6,7 +6,7 @@ import 'package:places/shared/places_texts.dart';
 import 'package:places/ui/common/icons.dart';
 import 'package:places/ui/common/rainbow_button.dart';
 import 'package:places/ui/navigation/sight_bottom_navigator.dart';
-import 'package:places/ui/screen/add_sight_screen.dart';
+import 'package:places/ui/screen/add_sight_screen/add_sight_screen.dart';
 import 'package:places/ui/screen/cards/sight_list_card.dart';
 import 'package:places/ui/screen/sight_app_bar.dart';
 import 'package:places/ui/screen/widgets/search_bar.dart';
